requirements.txt
streamlit==1.49.1
langchain-groq==1.0.0
langchain-ollama==0.3.10
langchain-text-splitters==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
python-dotenv==1.2.1
faiss-cpu==1.9.0.post1
pypdf==5.1.0
